[70197] Refactor import/export file upload
