+++ Controller/Adminhtml/Menu/ImportPost.php
@@ -10,7 +10,7 @@ use Magento\Framework\App\Action\HttpPostActionInterface;
 use Magento\Framework\Exception\ValidatorException;
 use Magento\Store\Model\StoreManagerInterface;
 use Psr\Log\LoggerInterface;
-use Snowdog\Menu\Model\ImportExport\FileUpload;
+use Snowdog\Menu\Model\ImportExport\File\Upload as FileUpload;
 use Snowdog\Menu\Model\ImportExport\ImportProcessor;
 use Snowdog\Menu\Model\ImportExport\Processor\Import\Validator\ValidationAggregateError;
